enhance: split `redirects` file in two so we can properly mock it
